move linalg funcs to linalg module
