+++ statsmodels/regression/mixed_linear_model.py
@@ -158,53 +158,8 @@ import warnings
 from statsmodels.tools.sm_exceptions import ConvergenceWarning
 from statsmodels.base._penalties import Penalty
 from statsmodels.compat.numpy import np_matrix_rank
-
-
-def _dot(x, y):
-    """
-    Returns the dot product of the arrays, works for sparse and dense.
-    """
-
-    if isinstance(x, np.ndarray) and isinstance(y, np.ndarray):
-        return np.dot(x, y)
-    elif sparse.issparse(x):
-        return x.dot(y)
-    elif sparse.issparse(y):
-        return y.T.dot(x.T).T
-
-
-# From numpy, adapted to work with sparse and dense arrays.
-def _multi_dot_three(A, B, C):
-    """
-    Find best ordering for three arrays and do the multiplication.
-
-    Doing in manually instead of using dynamic programing is
-    approximately 15 times faster.
-    """
-    # cost1 = cost((AB)C)
-    cost1 = (A.shape[0] * A.shape[1] * B.shape[1] +  # (AB)
-             A.shape[0] * B.shape[1] * C.shape[1])   # (--)C
-    # cost2 = cost((AB)C)
-    cost2 = (B.shape[0] * B.shape[1] * C.shape[1] +  # (BC)
-             A.shape[0] * A.shape[1] * C.shape[1])   # A(--)
-
-    if cost1 < cost2:
-        return _dot(_dot(A, B), C)
-    else:
-        return _dot(A, _dot(B, C))
-
-
-def _dotsum(x, y):
-    """
-    Returns sum(x * y), where '*' is the pointwise product, computed
-    efficiently for dense and sparse matrices.
-    """
-
-    if sparse.issparse(x):
-        return x.multiply(y).sum()
-    else:
-        # This way usually avoids allocating a temporary.
-        return np.dot(x.ravel(), y.ravel())
+from statsmodels.tools.linalg import (_dot, _multi_dot_three, _dotsum,
+                                      _smw_solver, _smw_logdet)
 
 
 def _get_exog_re_names(self, exog_re):
@@ -411,98 +366,6 @@ class MixedLMParams(object):
             pa = np.concatenate((cpa, vcomp))
 
         return pa
-
-
-def _smw_solver(s, A, AtA, BI, di):
-    """
-    Solves the system (s*I + A*B*A') * x = rhs for an arbitrary rhs.
-
-    The inverse matrix of B is block diagonal.  The upper left block
-    is BI and the lower right block is a diagonal matrix containing
-    di.
-
-    Parameters
-    ----------
-    s : scalar
-        See above for usage
-    A : ndarray
-        See above for usage
-    AtA : square ndarray
-        A.T * A
-    BI : square symmetric ndarray
-        The inverse of `B`.
-    di : array-like
-
-    Returns
-    -------
-    A function that takes `rhs` as an input argument and returns a
-    solution to the linear system defined above.
-    """
-
-    # Use SMW identity
-    qmat = AtA / s
-    m = BI.shape[0]
-    qmat[0:m, 0:m] += BI
-    ix = np.arange(m, A.shape[1])
-    qmat[ix, ix] += di
-    if sparse.issparse(A):
-        qi = sparse.linalg.inv(qmat)
-        qmati = A.dot(qi.T).T
-    else:
-        qmati = np.linalg.solve(qmat, A.T)
-
-    def solver(rhs):
-        if sparse.issparse(A):
-            ql = qmati.dot(rhs)
-            ql = A.dot(ql)
-        else:
-            ql = np.dot(qmati, rhs)
-            ql = np.dot(A, ql)
-        rslt = rhs / s - ql / s**2
-        if sparse.issparse(rslt):
-            rslt = np.asarray(rslt.todense())
-        return rslt
-
-    return solver
-
-
-def _smw_logdet(s, A, AtA, BI, di, B_logdet):
-    """
-    Returns the log determinant of s*I + A*B*A'.
-
-    Uses the matrix determinant lemma to accelerate the calculation.
-
-    Parameters
-    ----------
-    s : scalar
-        See above for usage
-    A : square symmetric ndarray
-        See above for usage
-    AtA : square matrix
-        A.T * A
-    BI : square symmetric ndarray
-        The upper left block of B^-1.
-    di : array-like
-        The diagonal elements of the lower right block of B^-1.
-    B_logdet : real
-        The log determinant of B
-
-    Returns
-    -------
-    The log determinant of s*I + A*B*A'.
-    """
-
-    p = A.shape[0]
-    ld = p * np.log(s)
-    qmat = AtA / s
-    m = BI.shape[0]
-    qmat[0:m, 0:m] += BI
-    ix = np.arange(m, A.shape[1])
-    qmat[ix, ix] += di
-    if sparse.issparse(qmat):
-        qmat = qmat.todense()
-    _, ld1 = np.linalg.slogdet(qmat)
-    return B_logdet + ld + ld1
 
 
 class MixedLM(base.LikelihoodModel):
